# src/main/resources/static/swagger-ui/openapi3.yaml
openapi: 3.0.1
info:
  title: My API
  description: My API description
  version: 0.1.0
servers:
- url: http://localhost:8080
...
paths:
...
  /api/v1/reviews/{id}:
    get:
      tags:
      - api
      summary: Retrieve reviews for a specific toilet by its ID.
      description: Retrieve reviews for a specific toilet by its ID.
      operationId: get-toilet-reviews
      parameters:
      - name: id
        in: path
        description: Unique identifier of the toilet.
        required: true
        schema:
          type: string
      - name: size
        in: query
        description: Number of reviews to retrieve per page. Default is 10.
        required: false
        schema:
          type: string
      - name: index
        in: query
        description: Page index for pagination. Default is 0.
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api-v1-reviews-id181444805"
              examples:
                get-toilet-reviews:
                  value: "[{\"stars\":2,\"comment\":\"생각보다 깔끔해요\",\"cleanliness\"\
                    :false,\"toiletPaper\":true,\"waterPressure\":false,\"hotWater\"\
                    :false,\"safety\":true,\"createdAt\":\"2024-01-22T18:00:00Z\"\
                    },{\"stars\":5,\"comment\":\"안전한 편이에요\",\"cleanliness\":true,\"\
                    toiletPaper\":false,\"waterPressure\":true,\"hotWater\":true,\"\
                    safety\":false,\"createdAt\":\"2024-01-21T09:00:00Z\"},{\"stars\"\
                    :2,\"comment\":\"깨끗하고 좋아요\",\"cleanliness\":false,\"toiletPaper\"\
                    :true,\"waterPressure\":false,\"hotWater\":true,\"safety\":true,\"\
                    createdAt\":\"2024-01-20T22:00:00Z\"},{\"stars\":5,\"comment\"\
                    :\"깨끗하고 좋아요\",\"cleanliness\":true,\"toiletPaper\":false,\"waterPressure\"\
                    :false,\"hotWater\":true,\"safety\":true,\"createdAt\":\"2024-01-19T20:00:00Z\"\
                    },{\"stars\":4,\"comment\":\"수압이 약해요\",\"cleanliness\":true,\"\
                    toiletPaper\":true,\"waterPressure\":true,\"hotWater\":true,\"\
                    safety\":false,\"createdAt\":\"2024-01-18T04:00:00Z\"},{\"stars\"\
                    :2,\"comment\":null,\"cleanliness\":true,\"toiletPaper\":false,\"\
                    waterPressure\":true,\"hotWater\":false,\"safety\":false,\"createdAt\"\
                    :\"2024-01-17T14:00:00Z\"},{\"stars\":5,\"comment\":\"냄새가 좀 납니\
                    다\",\"cleanliness\":false,\"toiletPaper\":false,\"waterPressure\"\
                    :false,\"hotWater\":false,\"safety\":false,\"createdAt\":\"2024-01-16T15:00:00Z\"\
                    },{\"stars\":4,\"comment\":\"깨끗하고 좋아요\",\"cleanliness\":false,\"\
                    toiletPaper\":false,\"waterPressure\":true,\"hotWater\":true,\"\
                    safety\":false,\"createdAt\":\"2024-01-15T16:00:00Z\"},{\"stars\"\
                    :3,\"comment\":\"화장지가 없어요\",\"cleanliness\":false,\"toiletPaper\"\
                    :true,\"waterPressure\":false,\"hotWater\":true,\"safety\":false,\"\
                    createdAt\":\"2024-01-14T08:00:00Z\"},{\"stars\":3,\"comment\"\
                    :\"관리가 잘 되어 있어요\",\"cleanliness\":true,\"toiletPaper\":true,\"\
                    waterPressure\":false,\"hotWater\":false,\"safety\":false,\"createdAt\"\
                    :\"2024-01-13T10:00:00Z\"}]"
    post:
      tags:
      - api
      summary: Submit a review for a specific toilet by its ID.
      description: Submit a review for a specific toilet by its ID.
      operationId: post-toilet-review
      parameters:
      - name: id
        in: path
        description: Unique identifier of the toilet.
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/api-v1-reviews-id973051148"
            examples:
              post-toilet-review:
                value: "{\"stars\":2,\"comment\":\"硤त⨑蚝睋Ჯꎼᶜ퉤䑀⋮\",\"likes\":{\"cleanliness\"\
                  :false,\"toiletPaper\":false,\"waterPressure\":false,\"hotWater\"\
                  :false,\"safety\":true}}"
      responses:
        "200":
          description: "200"
...
  /api/v1/reviews/{id}/rating:
    get:
      tags:
      - api
      summary: Retrieve rating information about a specific toilet by its ID.
      description: Retrieve rating information about a specific toilet by its ID.
      operationId: get-toilet-rating
      parameters:
      - name: id
        in: path
        description: Unique identifier of the toilet.
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api-v1-reviews-id-rating-2058800511"
              examples:
                get-toilet-rating:
                  value: "{\"avgStars\":3.1,\"avgCleanliness\":2.3,\"avgToiletPaper\"\
                    :2.5,\"avgWaterPressure\":2.4,\"avgHotWater\":2.5,\"avgSafety\"\
                    :2.3}"
components:
  schemas:
    api-v1-reviews-id973051148:
      type: object
      properties:
        comment:
          type: string
          description: Optional comment about the toilet.
        stars:
          type: number
          description: Star rating for the toilet (1 to 5).
        likes:
          type: object
          properties:
            safety:
              type: boolean
              description: Indicates if the reviewer felt safe using the toilet.
            waterPressure:
              type: boolean
              description: Indicates if the reviewer liked the water pressure.
            toiletPaper:
              type: boolean
              description: Indicates if the reviewer liked the availability of toilet
                paper.
            hotWater:
              type: boolean
              description: Indicates if the reviewer liked the availability of hot
                water.
            cleanliness:
              type: boolean
              description: Indicates if the reviewer liked the cleanliness of the
                toilet.
    api-v1-reviews-id181444805:
      type: array
      description: List of items
      items:
        type: object
        properties:
          createdAt:
            type: string
            description: Timestamp when the review was created.
          safety:
            type: boolean
            description: Indicates if safety was liked.
          waterPressure:
            type: boolean
            description: Indicates if water pressure was liked.
          toiletPaper:
            type: boolean
            description: Indicates if toilet paper availability was liked.
          comment:
            type: string
            description: Comment provided in the review.
          stars:
            type: number
            description: Star rating given in the review.
          hotWater:
            type: boolean
            description: Indicates if hot water availability was liked.
          cleanliness:
            type: boolean
            description: Indicates if cleanliness was liked.
        description: List of items
...
    api-v1-reviews-id-rating-2058800511:
      type: object
      properties:
        avgHotWater:
          type: number
          description: Average hot water availability rating.
        avgStars:
          type: number
          description: Average star rating of the toilet.
        avgSafety:
          type: number
          description: Average safety rating.
        avgToiletPaper:
          type: number
          description: Average toilet paper availability rating.
        avgWaterPressure:
          type: number
          description: Average water pressure rating.
        avgCleanliness:
          type: number
          description: Average cleanliness rating.
...